update and create updates

+++ api/quotes/delete.php
@@ -5,6 +5,11 @@
 
         }
         public function delete($data){
+            
+            if(!array_key_exists('id', $data)){
+                echo json_encode(["message" => 'Missing Required Parameters']);
+                exit;
+            }
 
             
             if($this->gateway->delete($data)){
